python/requirements.txt
eth-brownie
pandas
google-cloud-bigquery 
ipython 
